subscriber/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["kenchin110100 <11845307+kenchin110100@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.1"
urllib3 = ">=1.21.1,<1.27"
gunicorn = "^20.1.0"
llama-index = "0.6.1"
pinecone-client = "^2.2.1"
transformers = "^4.28.1"
html2text = "^2020.1.16"
google-cloud-secret-manager = "^2.16.1"
sqlalchemy = "2.0.12"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^1.2.0"
black = "^23.3.0"
isort = "^5.12.0"
pre-commit = "^3.2.2"
flake8-pyproject = "^1.2.3"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = true

[tool.flake8]
max-line-length = 120
ignore = "E203,W503"
exclude = [".venv", ".git", "__pycache__",]
max-complexity = 10

[tool.mypy]
files = ["src/*.py"]